Refactor upgrade_helpers.py
